fix: bugs

+++ nimonspoli-gui/src/core/Card/ElectionCard.cpp
@@ -22,7 +22,7 @@ void ElectionCard::execute(Player &p, GameState &gs)
     {
         if (other == &p)
             continue;
-        p -= (int)amountPerPlayer;
-        *other += (int)amountPerPlayer;
+        p -= amountPerPlayer;
+        *other += amountPerPlayer;
     }
 }
